+++ sphinx/source/changelog.rst
@@ -3,6 +3,8 @@ Changelog
 
 5.4b1
 -----
+* Added the :ref:`scr2img.py` command (for converting the screenshot in a SCR
+  file or SNA/SZX/Z80 snapshot into a PNG or GIF file)
 * Added the ``--variables`` option to :ref:`snapinfo.py` (for showing the
   contents of the variables area)